requirements.txt
decorator==5.0.9
distlib==0.3.1
Django==3.2
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
Pygments==2.11.2
PyJWT==1.7.1
python3-openid==3.2.0
PyYAML==6.0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.2
sqlparse==0.4.2
static3==0.7.0
tablib==3.1.0
text-unidecode==1.2
urllib3==1.26.7
uWSGI==2.0.20
virtualenv==20.4.3
w3lib==1.22.0
whitenoise==5.3.0